Isolate SearchItem preview.
